@@ -3,7 +3,7 @@ package com.originalnexus.mediator.dialogs;
 import android.app.Dialog;
 import android.content.DialogInterface;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.app.DialogFragment;
 import android.app.AlertDialog;
 
