@@ -2,7 +2,7 @@ package com.originalnexus.mediator.dialogs;
 
 import android.app.AlertDialog;
 import android.app.Dialog;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.app.DialogFragment;
 import android.content.DialogInterface;
 import android.os.Bundle;
