@@ -1,6 +1,6 @@
 package com.originalnexus.mediator.models;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import com.google.gson.annotations.SerializedName;
 
